Separate crash simulation and breakpad wrappers
