+++ jni/breakpad.h
@@ -1,0 +1,6 @@
+#ifndef __BREAKPAD_H
+#define __BREAKPAD_H
+
+void setUpBreakpad(const char* path);
+
+#endif /*__BREAKPAD_H*/
