+++ jni/crash.h
@@ -1,0 +1,7 @@
+#ifndef __CRASH_H
+#define __CRASH_H
+
+void CrashWithStack();
+void CrashWithHeap();
+
+#endif /*__CRASH_H*/
